+++ mackup/main.py
@@ -5,7 +5,7 @@ Copyright (C) 2013-2021 Laurent Raufaste <http://glop.org/>
 
 Usage:
   mackup list
-  mackup [options] backup
+  mackup [options] symlink
   mackup [options] restore
   mackup show <application>
   mackup [options] uninstall
@@ -22,7 +22,7 @@ Options:
 
 Modes of action:
  1. list: display a list of all supported applications.
- 2. backup: sync your conf files to your synced storage, use this the 1st time
+ 2. symlink: sync your conf files to your synced storage, use this the 1st time
     you use Mackup.
  3. restore: link the conf files already in your synced storage on your system,
     use it on any new system you use.
@@ -83,9 +83,9 @@ def main():
 
     verbose = args["--verbose"]
 
-    if args["backup"]:
+    if args["symlink"]:
         # Check the env where the command is being run
-        mckp.check_for_usable_backup_env()
+        mckp.check_for_usable_symlink_env()
 
         if utils.confirm(
             "DANGEROUS: Mackup will REPLACE your original configuration files with symlinks to the ones in the "
@@ -100,11 +100,11 @@ def main():
                 "-really-do-to-my-files before you take any further actions.\n"
                 "Do you sure you want to continue?"
             ):
-                # Backup each application
-                for app_name in sorted(mckp.get_apps_to_backup()):
+                # Symlink each application
+                for app_name in sorted(mckp.get_apps_to_symlink()):
                     app = ApplicationProfile(mckp, app_db.get_files(app_name), dry_run, verbose)
                     printAppHeader(app_name)
-                    app.backup()
+                    app.symlink()
             else:
                 utils.error("It seems you are not sure what you are doing. Exiting...")
         else:
@@ -128,7 +128,7 @@ def main():
         app_db = ApplicationsDatabase()
 
         # Restore the rest of the app configs, using the restored Mackup config
-        app_names = mckp.get_apps_to_backup()
+        app_names = mckp.get_apps_to_symlink()
         # Mackup has already been done
         app_names.discard(MACKUP_APP_NAME)
 
@@ -152,7 +152,7 @@ def main():
         ):
             # Uninstall the apps except Mackup, which we'll uninstall last, to
             # keep the settings as long as possible
-            app_names = mckp.get_apps_to_backup()
+            app_names = mckp.get_apps_to_symlink()
             app_names.discard(MACKUP_APP_NAME)
 
             for app_name in sorted(app_names):